+++ tests/testthat/test-PKNCA.options.R
@@ -295,6 +295,25 @@ test_that("PKNCA.choose.option", {
                                      foo="bar",
                                      max.aucinf.pext=10)),
                10)
+  
+  ## Manage NULL and the "value" argument
+  expect_equal(PKNCA.choose.option("single.dose.aucs"),
+               PKNCA.options(name="single.dose.aucs"),
+               info="PKNCA.choose.option gives the default option when nothing else is given")
+  expect_equal(PKNCA.choose.option("single.dose.aucs",
+                                   options=list(single.dose.aucs=data.frame(start=0, end=1, cmax=TRUE))),
+               check.interval.specification(data.frame(start=0, end=1, cmax=TRUE)),
+               info="PKNCA.choose.option gives the default option when nothing else is given")
+  expect_equal(PKNCA.choose.option("single.dose.aucs",
+                                   value=data.frame(start=0, end=1, cmax=TRUE, tmax=TRUE),
+                                   options=list(single.dose.aucs=data.frame(start=0, end=1, cmax=TRUE))),
+               check.interval.specification(data.frame(start=0, end=1, cmax=TRUE, tmax=TRUE)),
+               info="PKNCA.choose.option gives the default option when nothing else is given")
+  expect_equal(PKNCA.choose.option("single.dose.aucs",
+                                   value=NULL,
+                                   options=list(single.dose.aucs=data.frame(start=0, end=1, cmax=TRUE))),
+               check.interval.specification(data.frame(start=0, end=1, cmax=TRUE)),
+               info="PKNCA.choose.option gives the default option when nothing else is given")
 })
 
 context("PKNCA summary setting")