+++ tests/testthat/test-pk.calc.all.R
@@ -361,3 +361,43 @@ test_that("Missing dose info for some subjects gives a warning, not a difficult-
                 all(myresult$result[["PPTESTCD"]] == rep(c("auclast", "cl.last"), 2)),
               info="cl.last is not calculated when dose information is missing, but only for the subject where dose info is missing.")
 })
+
+# Fix issue #68
+test_that("Ensure that options are respected during pk.nca call", {
+  doses <- data.frame(ID=1:2, Time=0, Dose=0.5)
+  
+  conc.data <- c(0, 1, 2, 1.3, 0.4, 0.35, 0.125)
+  time.data <- c(0, 1, 2, 4,   8,   24,   48)
+  concs <- merge(doses[c("ID")], data.frame(Conc=conc.data, Time=time.data))
+  
+  myconc <- PKNCA::PKNCAconc(concs, formula=Conc~Time|ID)
+  mydose <- PKNCA::PKNCAdose(doses, formula=Dose~Time|ID)
+  
+  myintervals <- data.frame(start=c(0,0,0),
+                            end=c(24,48,Inf),
+                            auclast=TRUE,
+                            aucinf.obs=TRUE,
+                            aucinf.pred=TRUE,
+                            aumclast=TRUE,
+                            aumcall=TRUE,
+                            half.life=TRUE)
+  
+  linear.mydata <- PKNCA::PKNCAdata(myconc, mydose, intervals = myintervals,
+                                    options = list(auc.method = "linear"))
+  linear.results <- PKNCA::pk.nca(linear.mydata)
+  
+  linlog.mydata <- PKNCA::PKNCAdata(myconc, mydose, intervals = myintervals,
+                                    options = list(auc.method = "lin up/log down"))
+  linlog.results <- PKNCA::pk.nca(linlog.mydata)
+  expect_true(all.equal(linear.results$result$PPORRES[linear.results$result$PPTESTCD %in% "aucinf.obs" &
+                                                         linear.results$result$ID %in% 1 &
+                                                         linear.results$result$end %in% Inf],
+                         24.54319,
+                         tolerance=0.0001) &
+              all.equal(linlog.results$result$PPORRES[linlog.results$result$PPTESTCD %in% "aucinf.obs" &
+                                                        linlog.results$result$ID %in% 1 &
+                                                        linlog.results$result$end %in% Inf],
+                        23.68317,
+                        tolerance=0.0001),
+              info="linear and loglinear effects are calculated differently.")
+})